Moved icon picker to folder
