+++ src/components/common/icon-picker/iconPicker.scss
